+++ extension/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 3.3.0
+
+### Added
+
+- Upgrade excalidraw package to 1.2.0
+- Add the ability to configure the export options for `excalidraw.png` and `excalidraw.svg` files using the `excalidraw.image` setting.
+
 ## 3.2.0
 
 ### Fixed